# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/ai-study/status": {
      "get": {
        "operationId": "AiStudyController_status",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["AI 错题复盘"]
      }
    },
    "/api/ai-study/reports": {
      "get": {
        "operationId": "AiStudyController_list",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 10000,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["AI 错题复盘"]
      },
      "post": {
        "operationId": "AiStudyController_create",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["AI 错题复盘"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mistakeIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "minItems": 1,
                    "maxItems": 10
                  },
                  "reflection": {
                    "type": "string",
                    "maxLength": 2000,
                    "default": ""
                  }
                },
                "required": ["mistakeIds"],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/ai-study/reports/{id}": {
      "get": {
        "operationId": "AiStudyController_report",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["AI 错题复盘"]
      },
      "delete": {
        "operationId": "AiStudyController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["AI 错题复盘"]
      }
    },
    "/api/ai-study/reports/{id}/search": {
      "post": {
        "operationId": "AiStudyController_search",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["AI 错题复盘"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 300
                  }
                },
                "required": ["query"],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/ai-study/reports/{id}/export": {
      "get": {
        "operationId": "AiStudyController_export",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "schema": {
              "type": "string",
              "enum": ["md", "json"],
              "default": "md"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["AI 错题复盘"]
      }
    },
    "/api/ai-study/reports/{id}/sources/{sourceId}/download": {
      "post": {
        "operationId": "AiStudyController_download",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sourceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["AI 错题复盘"]
      }
    },
...
  },
  "info": {
    "title": "知学 · 学习管理系统 API",
    "description": "同源 Cookie 会话。写请求必须携带登录响应中的 x-csrf-token。所有教学资源逐请求执行机构、当前角色、课程成员资格和对象归属校验。分页 page/pageSize，错误 {error:{code,message,fields,requestId}}。详细参数及示例见 docs/*-api.md。",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lms_session"
      }
    },
...
  }
}